+++ ssMemoWeb/bundle.py
@@ -18,6 +18,7 @@ MODULE_ORDER = [
     'src/markdown.js',
     'src/file-utils.js',
     'src/encoding.js',
+    'src/calc.js',
     'src/data-sources/indexeddb.js',
     'src/data-sources/rest.js',
     'src/data-sources/memory.js',